requirements.txt
# === Deep Learning Frameworks ===
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3

# === YOLO ===
ultralytics>=8.0.0

# === Geospatial ===
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
gdal

# === Image Processing ===
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
albumentations>=1.3.0

# === Data & ML ===
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# === Utilities ===
pyyaml>=6.0
tensorboard>=2.14.0
